Fix error reporting at end of file
When the source text ends early (missing some required syntax),
attempting to report the line and column of the error results in
attempting to get line and column of one character past the end of the
terminals, resulting in an IndexError exception.  This commit contains
a simple fix for this situation.