Fix bug with inlining recursive rules
When parsing sequence items, the item parser uses the parent model
type for the sequence item parser and inlines the result.  We were
detecting this situation by checking that the item result type was the
same as the alternatives result type, which spuriously triggered when
the item was not a sequence but a recursive invocation of the same
rule.  This modifies this logic to not be driven by result type but by
an explicit inline_to_parent flag.