Fix a regression caused by all caps _TRIVIA
Just before the 0.1.0 release we renamed the trivia rule from "trivia"
to "_TRIVIA" to prevent potential conflicts with user-defined rules.
Unfortunately the FLTK grammar itself insists on lowercase identifiers,
and for a somewhat good reason; identifier names end up in the generated
parser function names, and linters don't like all caps in function
names. It's nicer on the eyes to use "_trivia" anyway, so we're
switching to that.