Fix linter errors in gencode
Due to pyproject.toml misconfiguration we were not running pyright on
the files in the repo root, which includes the fltk parser and CST
files. So we had some linter errors in the gencode that weren't being
detected. This fixes the misconfiguration so that we lint those, and
fixes the linter errors.